Authentication

+++ module_1/exercises_rest_express/the-good-corner/backend/src/graphql/resolvers/tag-resolver.ts
@@ -1,10 +1,11 @@
-import { Resolver, Query, Arg, Int, Mutation } from "type-graphql";
+import { Resolver, Query, Arg, Int, Mutation, Authorized, Ctx } from "type-graphql";
 import { Tag } from "../../database/entities/tag";
 import { AppError } from "../../middlewares/error-handler";
 import { CreateTagInput } from "../inputs/create/create-tag-input";
 import { generateRandomTag } from "../../utils/generate-tag";
 import { UpdateTagInput } from "../inputs/update/update-tag-input";
 import dataSource from '../../database/config/datasource';
+import { AuthContext } from "../../types/types";
 
 @Resolver(Tag)
 export class TagResolver {
@@ -30,8 +31,24 @@ export class TagResolver {
     }
 
     // Create a new tag
+    @Authorized()
     @Mutation(() => Tag)
-    async createTag(@Arg('data') data: CreateTagInput): Promise<Tag> {
+    async createTag(
+        @Arg('data') data: CreateTagInput,
+        @Ctx() context: AuthContext
+    ): Promise<Tag> {
+        const user = context.user; // Get the user from the context
+
+        // Check if the user is authenticated
+        if (!user) {
+            throw new AppError("User must be authenticated to create a tag", 401, "UnauthorizedError");
+        }
+
+        // Check if the user is an admin
+        if (user.role !== "admin") {
+            throw new AppError("You are not authorized to create a tag", 403, "ForbiddenError");
+        }
+
         const tagRepository = dataSource.getRepository(Tag); // Get the tag repository
 
         // Check if a tag with the same title already exists
@@ -41,7 +58,11 @@ export class TagResolver {
             throw new AppError('A tag with this title already exists', 400, 'ValidationError');
         }
 
-        const tag = tagRepository.create({ ...data }); // Create a new tag
+        const tag = tagRepository.create({
+            ...data,
+            createdBy: user,
+        }); // Create a new tag
+
         return await tagRepository.save(tag); // Save the tag
     }
 
@@ -55,17 +76,35 @@ export class TagResolver {
     }
 
     // Update a tag partially (PATCH)
+    @Authorized()
     @Mutation(() => Tag)
     async updateTag(
         @Arg('id', () => Int) id: number, // The id of the tag to update
-        @Arg('data') data: UpdateTagInput // The data to update the tag
+        @Arg('data') data: UpdateTagInput, // The data to update the tag
+        @Ctx() context: AuthContext
     ): Promise<Tag> {
+        const user = context.user; // Get the user from the context
+
+        // Check if the user is authenticated
+        if (!user) {
+            throw new AppError("User must be authenticated to update a tag", 401, "UnauthorizedError");
+        }
+
+        // Check if the user is an admin
+        if (user.role !== "admin") {
+            throw new AppError("You are not authorized to update a tag", 403, "ForbiddenError");
+        }
+
         const tagRepository = dataSource.getRepository(Tag); // Get the tag repository
 
-        const tag = await tagRepository.findOne({ where: { id }, relations: ['ads'] }); // Find a tag by its id with its ads
+        const tag = await tagRepository.findOne({ where: { id }, relations: ['ads', 'createdBy'] }); // Find a tag by its id with its ads
 
         if (!tag) { // If the tag is not found, throw an error
             throw new AppError('Tag not found', 404, 'NotFoundError');
+        }
+
+        if (tag.createdBy.id !== user.id) {
+            throw new AppError("You are not authorized to update this tag", 403, "ForbiddenError");
         }
 
         Object.assign(tag, data); // Update the tag
@@ -73,9 +112,32 @@ export class TagResolver {
     }
 
     // Delete a tag
+    @Authorized()
     @Mutation(() => Boolean)
-    async deleteTag(@Arg('id', () => Int) id: number): Promise<boolean> {
+    async deleteTag(@Arg('id', () => Int) id: number, @Ctx() context: AuthContext): Promise<boolean> {
+        const user = context.user; // Get the user from the context
+
+        // Check if the user is authenticated
+        if (!user) {
+            throw new AppError("User must be authenticated to delete a tag", 401, "UnauthorizedError");
+        }
+
+        // Check if the user is an admin
+        if (user.role !== "admin") {
+            throw new AppError("You are not authorized to delete a tag", 403, "ForbiddenError");
+        }
+
         const tagRepository = dataSource.getRepository(Tag); // Get the tag repository
+
+        const tag = await tagRepository.findOne({ where: { id }, relations: ['createdBy'] }); // Find a tag by its id with its creator
+        if (!tag) {
+            throw new AppError('Tag not found', 404, 'NotFoundError');
+        }
+
+        // Check if the user is the creator
+        if (tag.createdBy.id !== user.id) {
+            throw new AppError("You are not authorized to delete this tag", 403, "ForbiddenError");
+        }
 
         const deleteResult = await tagRepository.delete(id); // Delete the tag
